Pull date range index into it's own class
